light_management_project/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
distlib==0.3.8
Django==3.2.25
djangorestframework==3.15.0
filelock==3.12.2
gunicorn==21.2.0
importlib-metadata==6.7.0
packaging==24.0
platformdirs==4.0.0
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.7.1
virtualenv==20.25.1
zipp==3.15.0
